.backup/-5.-02_20-58-53/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
astroid==3.3.9
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
baidusearch==1.0.3
beartype==0.12.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.4
botocore==1.38.4
browser-use==0.1.41
browsergym==0.13.3
browsergym-assistantbench==0.13.3
browsergym-core==0.13.3
browsergym-experiments==0.13.3
browsergym-miniwob==0.13.3
browsergym-visualwebarena==0.13.3
browsergym-webarena==0.13.3
browsergym-workarena==0.4.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.5.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
duckduckgo_search==8.0.1
english-words==2.0.1
evaluate==0.4.3
executing==2.2.0
faiss-cpu==1.11.0
Faker==37.1.0
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
gymnasium==1.1.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.38
lazy_loader==0.4
libvisualwebarena==0.0.15
libwebarena==0.0.4
loguru==0.7.3
lxml==5.4.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.6.0
mem0ai==0.1.88
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.0
ollama==0.4.8
openai==1.76.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
Pillow==10.1.0
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
pocketflow_framework==0.0.1
portalocker==2.10.1
posthog==3.25.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pylint==3.3.6
pyparsing==3.2.3
pyperclip==1.9.0
pyqtdarktheme==0.1.7
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pyvis==0.3.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
-e git+https://github.com/alexpdev/QStyler.git@3e7456068b0f6f80e0fc3911f7a2e29e05a585ea#egg=QStyler
-e git+https://github.com/ESSS/qt_style_sheet_inspector.git@2d6e9ae2378d125c4299f758e666b00cf81504af#egg=qt_style_sheet_inspector
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.12.0
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
sentence-transformers==4.1.0
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
text-generation==0.7.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-requests==2.32.0.20250328
types-tqdm==4.67.0.20250417
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unidiff==0.7.5
urllib3==2.4.0
uvicorn==0.34.2
vulture==2.14
wcwidth==0.2.13
weblinx==0.3.2
weblinx_browsergym==0.0.3
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
